+++ packages/docs/docs/licensing/index.mdx
@@ -1,0 +1,25 @@
+---
+sidebar_label: Overview
+title: '@remotion/licensing'
+---
+
+_available from v4.0.237_
+
+This package allows holders of the [company license](https://remotion.pro/license) to send events to Remotion to track the following usage of Remotion products:
+
+- Cloud renders
+- WebCodecs conversions
+
+Also, it offers an API to programmatically check the usage in order to implement spend controls.
+
+## Installation
+
+<Installation pkg="@remotion/licensing" />
+
+import {TableOfContents} from './TableOfContents';
+
+<TableOfContents />
+
+## license
+
+[Remotion License](https://remotion.dev/license)
