+++ packages/docs/package.json
@@ -46,6 +46,7 @@
 		"@remotion/install-whisper-cpp": "workspace:*",
 		"@remotion/lambda": "workspace:*",
 		"@remotion/layout-utils": "workspace:*",
+		"@remotion/licensing": "workspace:*",
 		"@remotion/lottie": "workspace:*",
 		"@remotion/media-utils": "workspace:*",
 		"@remotion/motion-blur": "workspace:*",
