+++ packages/licensing/src/index.ts
@@ -1,5 +1,5 @@
-import {registerCompanyLicenseUsageEvent} from './register-company-license-usage-point';
-
-export const LicensingInternals = {
-	trackWebCodecConversion: registerCompanyLicenseUsageEvent,
-};
+export {
+	UsageEventClassification as Classification,
+	RegisterUsageEventResponse,
+	registerUsageEvent,
+} from './register-usage-point';
