+++ packages/licensing/src/register-usage-point.ts
@@ -1,0 +1,63 @@
+const HOST = 'http://localhost:51748/';
+
+type ApiResponse =
+	| {
+			success: true;
+			billable: boolean;
+			classification: UsageEventClassification;
+	  }
+	| {
+			success: false;
+			error: string;
+	  };
+
+export type RegisterUsageEventResponse = {
+	success: true;
+	billable: boolean;
+	classification: UsageEventClassification;
+};
+
+type UsageEventType = 'webcodec-conversion' | 'cloud-render';
+
+export type UsageEventClassification = 'billable' | 'development' | 'failed';
+
+export const registerUsageEvent = async ({
+	apiKey,
+	host,
+	succeeded,
+	event,
+}: {
+	apiKey: string;
+	host: string;
+	succeeded: boolean;
+	event: UsageEventType;
+}): Promise<RegisterUsageEventResponse> => {
+	const res = await fetch(`${HOST}/api/track`, {
+		method: 'POST',
+		body: JSON.stringify({
+			event: event,
+			apiKey,
+			host,
+			succeeded,
+		}),
+		headers: {
+			'Content-Type': 'application/json',
+		},
+	});
+	const json = (await res.json()) as ApiResponse;
+
+	if (json.success) {
+		return {
+			billable: json.billable,
+			success: true,
+			classification: json.classification,
+		};
+	}
+
+	if (!res.ok) {
+		throw new Error(json.error);
+	}
+
+	const read = await res.json();
+	return read;
+};
