+++ packages/licensing/src/test/register.test.ts
@@ -2,10 +2,10 @@
 // of a Remotion team member with an instance of remotion.pro running locally
 
 import {expect, test} from 'bun:test';
-import {registerCompanyLicenseUsageEvent} from '../register-company-license-usage-point';
+import {registerUsageEvent} from '../register-usage-point';
 
 test('Should be able to track production usage', async () => {
-	const result = await registerCompanyLicenseUsageEvent({
+	const result = await registerUsageEvent({
 		apiKey: 'rm_pub_1dd7193534bbe72571b55bd43926654ddaaca3dd6f07772b',
 		host: 'http://localhost:50955',
 		succeeded: true,
@@ -19,7 +19,7 @@ test('Should be able to track production usage', async () => {
 });
 
 test('Should be able to track development usage', async () => {
-	const result = await registerCompanyLicenseUsageEvent({
+	const result = await registerUsageEvent({
 		apiKey: 'rm_pub_1dd7193534bbe72571b55bd43926654ddaaca3dd6f07772b',
 		host: 'https://remotion.dev',
 		succeeded: true,
@@ -33,7 +33,7 @@ test('Should be able to track development usage', async () => {
 });
 
 test('Should reject invalid API key', async () => {
-	const result = registerCompanyLicenseUsageEvent({
+	const result = registerUsageEvent({
 		apiKey: 'rm_pub_1dd719b',
 		host: 'http://localhost:50955',
 		succeeded: true,
@@ -43,7 +43,7 @@ test('Should reject invalid API key', async () => {
 });
 
 test('Should reject invalid secret API key', async () => {
-	const result = registerCompanyLicenseUsageEvent({
+	const result = registerUsageEvent({
 		apiKey: 'rm_sec_1dd719b',
 		host: 'http://localhost:50955',
 		succeeded: true,
