+++ packages/licensing/tsconfig.json
@@ -4,8 +4,7 @@
 		"rootDir": "src",
 		"outDir": "dist",
 		"jsx": "react-jsx",
-		"skipLibCheck": true,
-		"target": "ES2023"
+		"skipLibCheck": true
 	},
 	"include": ["src"],
 	"references": []
